--- config/dbPool.mjs
@@ -1,11 +1,0 @@
-import pg from 'pg';
-import * as dotenv from 'dotenv';
-dotenv.config();
-
-export const Pool = new pg.Pool({
-    user: process.env.DB_USER,
-    host: process.env.DB_HOST,
-    database: process.env.DB_NAME,
-    password: process.env.DB_PASSWORD,
-    port: process.env.DB_PORT,
-});
